Add flag to color resources to invert syntax colors
