+++ dnGREP.WPF/Themes/DarkBrushes.xaml
@@ -1,10 +1,12 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d">
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
+                    xmlns:sys="clr-namespace:System;assembly=mscorlib">
 
-    
-    <SolidColorBrush x:Key="SplitterBrush" Color="#FF999999"/>
+
     <SolidColorBrush x:Key="AvalonEdit.Foreground" Color="#FFF5F5F5"/>
     <SolidColorBrush x:Key="AvalonEdit.Background" Color="#FF1E1E1E"/>
+    <sys:Boolean x:Key="AvalonEdit.SyntaxColor.Invert">true</sys:Boolean>
+
     <SolidColorBrush x:Key="Match.Highlight.Foreground" Color="#FFFFFFFF"/>
     <SolidColorBrush x:Key="Match.Highlight.Background" Color="#FF267800"/>
     <SolidColorBrush x:Key="Match.Skip.Foreground" Color="#FFFFFFFF"/>
@@ -18,6 +20,7 @@
         <GradientStop Color="#FF303030" Offset="0.8" />
     </LinearGradientBrush>
     <SolidColorBrush x:Key="Dialog.Background" Color="#FF404040"/>
+    <SolidColorBrush x:Key="SplitterBrush" Color="#FF999999"/>
 
 
     <!-- Window Caption -->
@@ -49,9 +52,9 @@
     <SolidColorBrush x:Key="Menu.SubMenu.Item.Highlighted.Foreground" Color="#FFFFFFFF" />
     <SolidColorBrush x:Key="Menu.SubMenu.Item.Highlighted.Background" Color="#FF094771" />
     <SolidColorBrush x:Key="Menu.SubItem.Item.Highlighted.Border" Color="#FF0C629C"/>
-    <SolidColorBrush x:Key="Menu.SubMenu.Item.Disabled.Foreground" Color="#FF999999" /> 
+    <SolidColorBrush x:Key="Menu.SubMenu.Item.Disabled.Foreground" Color="#FF999999" />
     <SolidColorBrush x:Key="Menu.CheckMark.Background" Color="#FF1475A6" />
-    <SolidColorBrush x:Key="Menu.CheckMark.Border" Color="#FFD2D2E6" /> 
+    <SolidColorBrush x:Key="Menu.CheckMark.Border" Color="#FFD2D2E6" />
     <SolidColorBrush x:Key="Menu.CheckMark.Foreground" Color="#FFFFFFFF" />
     <SolidColorBrush x:Key="Menu.CheckMark.Disabled.Background" Color="Transparent" />
     <SolidColorBrush x:Key="Menu.CheckMark.Disabled.Border" Color="#FF999999" />
@@ -154,7 +157,7 @@
     <SolidColorBrush x:Key="ComboBox.Disabled.Editable.Button.Border" Color="Transparent" />
     <SolidColorBrush x:Key="ComboBox.DropDown.Background" Color="#FF525252"/>
     <SolidColorBrush x:Key="ComboBox.DropDown.Border" Color="#FF808080"/>
-    
+
     <!-- Scroll Bar -->
     <SolidColorBrush x:Key="ScrollBar.Static.Background" Color="#FF363638"/>
     <SolidColorBrush x:Key="ScrollBar.Static.Border" Color="#FF363638"/>
@@ -189,7 +192,7 @@
     <SolidColorBrush x:Key="ProgressBar.Background" Color="#FF707070"/>
     <SolidColorBrush x:Key="ProgressBar.Border" Color="#FF404040"/>
     <SolidColorBrush x:Key="ProgressBar.Progress" Color="#FF00BA22"/>
-    
+
     <!-- Tree View -->
     <SolidColorBrush x:Key="TreeView.Foreground" Color="#FFFFFFFF"/>
     <SolidColorBrush x:Key="TreeView.Background" Color="#FF252526"/>
